Relocated interfaces & additional styles
- Relocated component interfaces within separate files
- Added animations to correct names
