+++ src/components/Box/ComponentInterface.ts
@@ -1,0 +1,7 @@
+export interface BoxProps {
+  name: any
+  type: any
+  uid?: any
+  profilePicture: any
+  isDropped: boolean
+}
