+++ src/components/Dustbin/ComponentInterface.ts
@@ -1,0 +1,6 @@
+export interface DustbinProps {
+  accept: string[]
+  lastDroppedItem?: any
+  profile: any
+  onDrop: (item: any) => void
+}
